Relocated some types to a private header for chat.  Begin extending IOnlineChat with Mixer-specific functionality
